+++ packages/nextra/src/webpack-plugins/nextra-search.ts
